+++ src/aks-preview/HISTORY.rst
@@ -12,6 +12,16 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
+18.0.0b15
++++++++
+* Fixed NPE issue for `--acns-transit-encryption-type`.
+
+18.0.0b14
++++++++
+* Add preview flag for `az aks namespace`.
+* Add `--node-provisioning-default-pools` to the `az aks update` command.
+* Add `--node-provisioning-default-pools` to the `az aks create` command.
+
 18.0.0b13
 +++++++
 * Add option `--enable-http-proxy` to `az aks update`.