+++ src/vme/HISTORY.rst
@@ -10,3 +10,8 @@ Release History
 1.0.0b2
 ++++++
 * Add new 'az vme list' command to list all version managed extensions.
+
+1.0.0b3
+++++++
+* Update 'az vme list --output table' to show correct versions.
+* Wait for the bundle feature flag to fully propagate after enabling it.